Allow globs in the source path specefication.
Now you can us eg `fcp somehost:~/foo*.tiff .` and have all the globbed files
transferred over. Note that you can ONLY put globs in the basename (ie, not in
any of the parent directory specifications). `find` is used for the glob
matching. On my machine that just calls out to libc's fnmatch() function.

Now passing the files to be transferred to tar via stdin (with the -I arg)
because the glob could produce multiple matches. Also there is a check or two
in the result of the glob that don't really expect it to be multiple lines but
also don't seem to break if that is the case.